requirements.txt
Pillow>=10.0.0
pyperclip>=1.8.2
pyinstaller>=5.13.0
